UpdateTask added
Added an UpdateTask to check for possible Updates. When executed, the
Task checks on the repository's site to check if the version of the
Application equals the current version that was pushed there. If not,
the User will be sent to the releases site of the repository by clicking
the response Label