TipCalculator
- Added a Pane for the Staff Member Area in the TipCalculator. This Way,
the Scene in itself can be redesigned without changing the Code for
deleting and adding a Member to the Tip Calculation
- Deleting now works as intended, Padding is still an issue, since the
Padding for the last row is still in effect after deleting a Staff
Member. Everything else should be wprking fine.
- Changed the ScrollPane for the TipWindow. Now it contains only the new
GridPane instead of the whole scene, so when scrolling, the User should
still be able to edit other Fields and click (future) Buttons, without
having to scroll to that area.